chore/console-refactoring
